Change naming and folder structure
