+++ lib/shopifyHMAC/index.js
@@ -2,6 +2,17 @@
 
 const crypto = require('crypto');
 
+const extractHmacParameter = ({ query }) => {
+  if (!query.hmac) {
+    throw new Error(
+      'Could not extract HMAC parameter: no HMAC parameter found in query.'
+    );
+  }
+  return query.hmac;
+};
+
+exports.extractHmacParameter = extractHmacParameter;
+
 const constructHmacParameter = (req, options) =>
   calculateHmacParam({
     algorithm: options.getAlgorithm(),
@@ -9,6 +20,8 @@ const constructHmacParameter = (req, options) =>
     message: getMessage({ ...req }),
     digest: options.getDigest(),
   });
+
+exports.constructHmacParameter = constructHmacParameter;
 
 const getMessage = ({ query }) => {
   // console.log('query', query);
@@ -33,7 +46,7 @@ const getMessage = ({ query }) => {
     }, [])
     .join('&');
 
-//   console.log('message', message);
+  //   console.log('message', message);
 
   return message;
 };
@@ -44,5 +57,3 @@ const calculateHmacParam = ({ algorithm, sharedSecret, message, digest }) =>
     .update(message)
     .digest(digest)
     .toString();
-
-module.exports = constructHmacParameter;
