+++ lib/shopifyHMAC/index.test.js
@@ -1,7 +1,62 @@
 'use strict';
 
 const { test } = require('tap');
-const constructHmacParameter = require('./constructHmacParameter');
+const {extractHmacParameter, constructHmacParameter} = require('.');
+
+test('Extracts a Hmac parameter from a valid query object', async ({
+  equal,
+}) => {
+  const tests = [
+    {
+      description: 'HMAC Parameter extracted from valid query object',
+      input: {
+        query: {
+          code: '0907a61c0c8d55e99db179b68161bc00',
+          hmac:
+            '700e2dadb827fcc8609e9d5ce208b2e9cdaab9df07390d2cbca10d7c328fc4bf',
+          shop: 'some-shop.myshopify.com',
+          state: '0.6784241404160823',
+          timestamp: '1337178173',
+        },
+      },
+      expected:
+        '700e2dadb827fcc8609e9d5ce208b2e9cdaab9df07390d2cbca10d7c328fc4bf',
+    },
+  ];
+
+  tests.forEach(({ description, input, expected }) => {
+    equal(extractHmacParameter(input), expected, description);
+  });
+});
+
+test('Throws an error when HMAC parameter cannot be extracted', async ({
+  throws,
+}) => {
+  const tests = [
+    {
+      description: 'No HMAC parameter in query string returns an error',
+      input: {
+        query: {
+          param: 'value',
+          pet: 'dog',
+        },
+      },
+      expected: new Error(
+        'Could not extract HMAC parameter: no HMAC parameter found in query.'
+      ),
+    },
+  ];
+
+  tests.forEach(({ description, input, expected }) => {
+    throws(
+      function () {
+        extractHmacParameter(input);
+      },
+      expected,
+      description
+    );
+  });
+});
 
 const defaultOptions = {
   sharedSecret: null,
@@ -12,6 +67,7 @@ const defaultOptions = {
     throw new Error('No getDigest function provided.');
   },
 };
+
 const defaultSignature =
   '700e2dadb827fcc8609e9d5ce208b2e9cdaab9df07390d2cbca10d7c328fc4bf';
 const defaultRequest = {
@@ -24,13 +80,13 @@ const defaultRequest = {
   },
 };
 
-test('Constructs a Hmac signature parameter string from the incoming request query', async ({
+test('Constructs a HMAC signature parameter string from the incoming request query', async ({
   equal,
 }) => {
   const tests = [
     {
       description:
-        'Matching signature constructed when a valid request is provided',
+        'Matching HMAC signature parameter constructed when a valid request is provided',
       input: defaultRequest,
       options: {
         ...defaultOptions,
@@ -42,7 +98,7 @@ test('Constructs a Hmac signature parameter string from the incoming request que
     },
     {
       description:
-        'Matching signature constructed when a valid request containing an ids array parameter is provided',
+        'Matching HMAC signature parameter constructed when a valid request containing an ids array parameter is provided',
       input: {
         query: {
           code: '0907a61c0c8d55e99db179b68161bc00',
@@ -51,7 +107,7 @@ test('Constructs a Hmac signature parameter string from the incoming request que
           shop: 'some-shop.myshopify.com',
           state: '0.6784241404160823',
           timestamp: '1337178173',
-          'ids[]': ['1', '2']
+          'ids[]': ['1', '2'],
         },
       },
       options: {
@@ -60,7 +116,8 @@ test('Constructs a Hmac signature parameter string from the incoming request que
         getAlgorithm: () => 'sha256',
         getDigest: () => 'hex',
       },
-      expected: '089cd381a3cbe82ba96e75d3244ec036621c4dfcb13285facd686bc4a7ed95dd',
+      expected:
+        '089cd381a3cbe82ba96e75d3244ec036621c4dfcb13285facd686bc4a7ed95dd',
     },
   ];
 
@@ -69,7 +126,7 @@ test('Constructs a Hmac signature parameter string from the incoming request que
   });
 });
 
-test('Throws an error when a Hmac parameter signature string cannot be constructed', async ({
+test('Throws an error when a HMAC signature parameter string cannot be constructed', async ({
   throws,
 }) => {
   const tests = [
