--- lib/shopifyHmac/extractHmacParameter.js
@@ -1,12 +1,0 @@
-"use strict";
-
-function extractHmacParameter({ query }) {
-  if (!query.hmac) {
-    throw new Error(
-      "Could not extract Hmac parameter: no Hmac parameter found in query."
-    );
-  }
-  return query.hmac;
-}
-
-module.exports = extractHmacParameter;
